Created Scores class and changed the type of some variables in the Player class to better fit the intended purpouse

+++ Tetris++/Tetris++/Player.cpp
@@ -3,7 +3,7 @@
 
 uint16_t Player::m_highScore;
 uint16_t Player::m_averageScore;
-int Player::m_scoreCount;
+uint8_t Player::m_scoreCount;
 
 Player::Player(std::string name, uint16_t score)
 	:m_name(name), m_score(score)
